Add files via upload
Added an extra Skill bar (SQL)

Fixed the whitespace that occured on mobile devices when scrolling right

Added the date and time in the navbar. Readying code to change colour depending on weather or date

Edited About me and Biography. 

Took away margin pushing the email box to the right side of the screen in mobile view.

Cleaning up and better organising code across multiple files.